+++ CHANGELOG.md
@@ -13,6 +13,9 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 - Added timeout for "waiting for pod to be ready..." in mirrord-layer to prevent unresponsive behavior. See [#579](https://github.com/metalbear-co/mirrord/issues/579)
 - IntelliJ Extension: Default log level to `ERROR` from `DEBUG`
 
+### Fixed
+- Issue with [bottlerocket](https://github.com/bottlerocket-os/bottlerocket) where they use `/run/dockershim.sock` instead of the default containerd path. Add new path as fallback.
+
 ## 3.2.0
 
 ### Changed